create evalScript function
